Toogle Button separated into own folder
